scrapers/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==5.1.0
playwright==1.42.0
httpx==0.27.0
python-dateutil==2.9.0
tqdm==4.66.2
python-dotenv==1.0.1
pydantic==2.6.4
aiohttp==3.9.3
fake-useragent==1.5.1
